Move flavours/glitch/utils/ready to flavours/glitch/ready
