+++ app/javascript/flavours/glitch/ready.js
