+++ app/javascript/flavours/glitch/utils/config.js
@@ -1,4 +1,4 @@
-import ready from './ready';
+import ready from '../ready';
 
 export let assetHost = '';
