+++ app/javascript/flavours/glitch/utils/main.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { setupBrowserNotifications } from 'flavours/glitch/actions/notifications';
 import Mastodon, { store } from 'flavours/glitch/containers/mastodon';
-import ready from 'flavours/glitch/utils/ready';
+import ready from 'flavours/glitch/ready';
 
 const perf = require('flavours/glitch/utils/performance');
